refactor: 도메인 클래스들 domain 패키지로 분리 및 LadderGameData rewards 변수 final 처리
