+++ src/main/java/ladder/domain/exception/InvalidCountOfPersonException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class InvalidCountOfPersonException extends RuntimeException {
