+++ src/main/java/ladder/domain/exception/InvalidLinesException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class InvalidLinesException extends RuntimeException {
