+++ src/main/java/ladder/domain/exception/InvalidNameLengthException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class InvalidNameLengthException extends RuntimeException {
