+++ src/main/java/ladder/domain/exception/InvalidRewordsException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class InvalidRewordsException extends RuntimeException {
