+++ src/main/java/ladder/domain/exception/NullNamesException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class NullNamesException extends RuntimeException {
