+++ src/main/java/ladder/domain/exception/UnknownNameException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.domain.exception;
 
 public class UnknownNameException extends RuntimeException {
