+++ src/main/java/ladder/domain/line/RandomLine.java
@@ -1,4 +1,4 @@
-package ladder.line;
+package ladder.domain.line;
 
 import java.util.Random;
