+++ src/test/java/ladder/domain/RewardsTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.ArrayList;
 import java.util.List;
-import ladder.exception.InvalidRewordsException;
+import ladder.domain.exception.InvalidRewordsException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
