+++ src/test/java/ladder/domain/line/ManualLine.java
@@ -1,4 +1,4 @@
-package ladder.line;
+package ladder.domain.line;
 
 import java.util.List;
